apriltag_ros/scripts/requirements.txt
numpy
pandas
pyaml
rosbags
scipy